--- x/tradebin/keeper/msg_server_test.go
@@ -1,16 +1,0 @@
-package keeper_test
-
-import (
-	"context"
-	"testing"
-
-	keepertest "github.com/bze-alphateam/bze/testutil/keeper"
-	"github.com/bze-alphateam/bze/x/tradebin/keeper"
-	"github.com/bze-alphateam/bze/x/tradebin/types"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-)
-
-func setupMsgServer(t testing.TB) (types.MsgServer, context.Context) {
-	k, ctx := keepertest.TradebinKeeper(t)
-	return keeper.NewMsgServerImpl(*k), sdk.WrapSDKContext(ctx)
-}
